refactor: imports posposet -> poset
