+++ src/cmd/db.go
@@ -10,17 +10,16 @@ import (
 	"github.com/Fantom-foundation/go-lachesis/src/internal"
 	"github.com/Fantom-foundation/go-lachesis/src/kvdb"
 	"github.com/Fantom-foundation/go-lachesis/src/metrics"
-	"github.com/Fantom-foundation/go-lachesis/src/posposet"
 )
 
-func makeStorages(makeDb internal.DbProducer) (*gossip.Store, *posposet.Store) {
+func makeStorages(makeDb internal.DbProducer) (*gossip.Store, *poset.Store) {
 	db := makeDb("lachesis")
 
 	g := db.NewTable([]byte("g_"))
 	p := db.NewTable([]byte("p_"))
 
 	return gossip.NewStore(g),
-		posposet.NewStore(p, makeDb)
+		poset.NewStore(p, makeDb)
 }
 
 func dbProducer(dbdir string) internal.DbProducer {
