+++ src/internal/integration.go
@@ -5,7 +5,7 @@ import (
 	"github.com/ethereum/go-ethereum/p2p/simulations/adapters"
 
 	"github.com/Fantom-foundation/go-lachesis/src/gossip"
-	"github.com/Fantom-foundation/go-lachesis/src/posposet"
+	"github.com/Fantom-foundation/go-lachesis/src/poset"
 )
 
 func NewIntegration(cfg *adapters.NodeConfig, config *gossip.Config) *gossip.Service {
